Continuing refactoring with correct data types
