+++ architecture-documentation/pom.xml
@@ -45,18 +45,15 @@
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
 			<artifactId>junit-jupiter-api</artifactId>
-			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
 			<artifactId>junit-jupiter-engine</artifactId>
-			<scope>test</scope>
 		</dependency>
 		<!-- https://mvnrepository.com/artifact/org.jboss.weld/weld-junit-parent -->
 		<dependency>
 			<groupId>org.jboss.weld</groupId>
-			<artifactId>weld-junit-parent</artifactId>
-			<type>pom</type>
+			<artifactId>weld-junit5</artifactId>
 		</dependency>
 	</dependencies>
