+++ maven-metadata-inferer/src/main/java/org/ndx/aadarchi/inferer/maven/enhancers/ComponentEnhancer.java
@@ -1,15 +1,15 @@
 package org.ndx.aadarchi.inferer.maven.enhancers;
 
 import org.apache.maven.project.MavenProject;
-import org.ndx.aadarchi.inferer.maven.MavenDetailsInfererEnhancer;
 import org.ndx.aadarchi.inferer.maven.MavenPomDecorator;
+import org.ndx.aadarchi.inferer.maven.MavenPomReader;
 
 import com.structurizr.model.Component;
 
 public class ComponentEnhancer extends ModelElementMavenEnhancer<Component> {
 
-	public ComponentEnhancer(MavenDetailsInfererEnhancer mavenDetailsInfererEnhancer, Component enhanced) {
-		super(mavenDetailsInfererEnhancer, enhanced);
+	public ComponentEnhancer(MavenPomReader mavenPomReader, Component enhanced) {
+		super(enhanced);
 	}
 
 	@Override
